CATS_v2/requirements.txt
# CATS v2.0 - Enhanced Evaluation Pipeline Requirements

# Core dependencies
anthropic>=0.39.0
httpx>=0.27.0
openai>=1.0.0
python-dotenv>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# NLP utilities
nltk>=3.8.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Machine Learning / Transformers (for local NLI models if needed)
torch>=2.0.0
transformers>=4.35.0

# Async & concurrency
aiohttp>=3.9.0
asyncio-throttle>=1.0.1

# Progress bars & logging
tqdm>=4.65.0
coloredlogs>=15.0.1

# Configuration management
pydantic>=2.0.0
pyyaml>=6.0

# Utilities
tenacity>=8.2.0
orjson>=3.9.0
